Title Screen Setup
Stole the title screen I made for Aloft and set it up over here. Simple panels and a slideout volume bar.

Also brought over the script that forces you to load in from a screen, and the nice button prefabs. Prefabs should be polished and probably reduced down to 1 button, and the script /probably/ isn't necessary for a singleplayer game, but I wanted to have it anyways